refactor(app): split standalone chat transport
Move Pi message preparation and steering flow out of the send transport hook while preserving the existing command surface.

Extract document folding, large-context preparation, image conversion, and prompt history wrapping into a focused preparation helper.

Move composer and queued steering behavior into a dedicated steering transport hook.

Tighten foreground Pi event parsing by narrowing unknown payloads before dispatch.

Keep chat props, event names, queue behavior, and Pi session semantics unchanged.
